+++ app/src/main/java/com/mahaperivaya/SendRequest/SendLogout.java
@@ -10,7 +10,7 @@ public class SendLogout {
   public String emailid;
 
   @SerializedName("profileid")
-  public int profileid;
+  public long profileid;
 
 
 }
